refactored name and updated groupId
