+++ build.gradle
@@ -3,7 +3,7 @@ plugins {
     id 'org.jetbrains.intellij' version '0.4.16'
 }
 
-group 'Open-Liberty'
+group 'io.openliberty.tools'
 version '0.0.1'
 
 sourceCompatibility = 1.8
@@ -13,7 +13,7 @@ repositories {
 }
 
 dependencies {
-    testCompile group: 'junit', name: 'junit', version: '4.12'
+    testImplementation group: 'junit', name: 'junit', version: '4.12'
 }
 
 runIde {
