+++ src/main/java/io/openliberty/tools/intellij/LibertyActionNode.java
@@ -1,4 +1,4 @@
-package org.liberty.intellij;
+package io.openliberty.tools.intellij;
 
 import javax.swing.tree.DefaultMutableTreeNode;
