+++ src/main/java/io/openliberty/tools/intellij/util/LibertyActionUtil.java
@@ -1,4 +1,4 @@
-package org.liberty.intellij.util;
+package io.openliberty.tools.intellij.util;
 
 import com.intellij.openapi.diagnostic.Logger;
 import com.jediterm.terminal.TtyConnector;
